fix(contracts): convert move and action phases to uint32

+++ packages/contracts/src/libraries/DSTypes.sol
@@ -40,6 +40,6 @@ struct GameConfig {
   // Block timestamp when the game started
   uint256 startTime;
   // Number of seconds
-  uint256 movePhaseLength;
-  uint256 actionPhaseLength;
+  uint32 movePhaseLength;
+  uint32 actionPhaseLength;
 }
